+++ cmd/paramount/main.go
@@ -14,8 +14,8 @@ func main() {
    flag.StringVar(&guid, "b", "", "GUID")
    // f
    // paramountplus.com/shows/video/186A7772-3307-4270-4FB9-D16F4DC7E08C
-   var bandwidth int
-   flag.IntVar(&bandwidth, "f", 2_223_000, "target bandwidth")
+   var bandwidth int64
+   flag.Int64Var(&bandwidth, "f", 2_223_000, "target bandwidth")
    // i
    var info bool
    flag.BoolVar(&info, "i", false, "information")
